+++ src/pages/MyPage/MyInfo.tsx
@@ -15,7 +15,7 @@ import { FiEdit2 } from 'react-icons/fi';
 import BottomNavigationBar from '@/components/common/BottomNavigationBar/BottomNavigationBar';
 import ConfirmModal from '@/components/modals/ConfirmModal';
 import { useState } from 'react';
-import { logout } from '@/api/auth';
+import { logout } from '@/api/auth/auth';
 import { useNavigate } from 'react-router-dom';
 
 export default function Myinfo() {
